learn/03-llm-apis/requirements.txt
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
numpy>=1.24.0
